Add files via upload
"Devilphishing" is a powerful phishing tool designed to aid penetration testers and security professionals in their efforts to test the strength of an organization's defenses against phishing attacks. The script utilizes advanced techniques to craft highly convincing phishing emails and websites, allowing users to easily simulate real-world phishing attacks.

One of the key features of Devilphishing is its ability to generate custom phishing pages for a wide range of popular websites, such as social media platforms, email providers, and financial institutions. These phishing pages are designed to mimic the look and feel of the real websites and are equipped with advanced features such as form field validation and working login forms, making them highly convincing to even the most discerning of users.

The script also includes a variety of customization options, allowing users to tailor the phishing attacks to their specific needs. This can include customizing the sender name and email address, as well as the subject and body of the phishing email. The script also supports the use of templates, which can be used to easily create new phishing pages based on pre-existing designs.

The script also includes detailed reporting capabilities, which can be used to track the success of phishing campaigns and identify potential vulnerabilities within an organization. Reports can be generated in a variety of formats, including CSV and HTML, and can be exported to other tools for further analysis.

Overall, Devilphishing is a powerful and highly customizable tool that can be used to test an organization's defenses against phishing attacks. It's ease of use and advanced features make it suitable for both novice and experienced users. However, it is important to remember that unauthorized phishing attempts are illegal, the script should be only used in the legal and ethical manner after gaining proper consent